requirements.txt
Authlib==1.2.0
autopep8==2.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
CacheControl
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
configparser==5.3.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
docopt==0.6.2
firebase-admin==6.1.0
Flask==2.3.2
fonttools==4.39.4
gcloud==0.17.0
google-api-core
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
houdini.py==0.1.0
html2text==2020.1.16
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jws==0.1.3
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
misaka==2.1.1
msgpack==1.0.5
numpy==1.21.1
oauth2client==3.0.0
packaging==23.1
Pillow==9.5.0
ply==3.11
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.7.0
pyparsing==3.0.9
Pyrebase
python-dateutil==2.8.2
python-jwt==2.0.1
PyYAML==6.0
requests
requests-toolbelt==0.7.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
threadpoolctl==3.1.0
toml.py==0.1.7
tomli==2.0.1
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.6
watchdog==3.0.0
Werkzeug==2.3.4
zipp==3.15.0